+++ gatsby-node.js
@@ -1,19 +1,7 @@
 const path = require('path')
 
-// Intersection Observer polyfill
-// requires `npm install intersection-observer`
-// https://github.com/gatsbyjs/gatsby/issues/2288#issuecomment-334467821
-exports.modifyWebpackConfig = ({ config, stage }) => {
-  if (stage === 'build-html') {
-    config.loader('null', {
-      test: /intersection-observer/,
-      loader: 'null-loader'
-    })
-  }
-}
-
-exports.createPages = ({ boundActionCreators, graphql }) => {
-  const { createPage } = boundActionCreators
+exports.createPages = ({ actions, graphql }) => {
+  const { createPage } = actions
 
   return new Promise((resolve, reject) => {
     const template = path.resolve('src/templates/Project.jsx')
@@ -32,7 +20,7 @@ exports.createPages = ({ boundActionCreators, graphql }) => {
                 img {
                   id
                   childImageSharp {
-                    sizes(maxWidth: 500, quality: 80) {
+                    fluid(maxWidth: 500, quality: 80) {
                       aspectRatio
                       src
                       srcSet
@@ -51,7 +39,7 @@ exports.createPages = ({ boundActionCreators, graphql }) => {
                 img {
                   id
                   childImageSharp {
-                    sizes(maxWidth: 500, quality: 80) {
+                    fluid(maxWidth: 500, quality: 80) {
                       aspectRatio
                       src
                       srcSet
