+++ package.json
@@ -21,8 +21,8 @@
   "dependencies": {
     "camel-case": "^3.0.0",
     "file-saver": "^1.3.8",
-    "gatsby": "^1.9.270",
-    "gatsby-image": "^1.0.52",
+    "gatsby": "next",
+    "gatsby-image": "next",
     "gatsby-link": "^1.6.44",
     "gatsby-plugin-favicon": "^2.1.1",
     "gatsby-plugin-matomo": "^0.4.0",
@@ -30,7 +30,7 @@
     "gatsby-plugin-react-helmet": "^2.0.11",
     "gatsby-plugin-react-next": "^1.0.11",
     "gatsby-plugin-sass": "^1.0.26",
-    "gatsby-plugin-sharp": "^1.6.47",
+    "gatsby-plugin-sharp": "next",
     "gatsby-plugin-sitemap": "^1.2.25",
     "gatsby-plugin-svgr": "^1.0.1",
     "gatsby-source-filesystem": "^1.5.38",
@@ -39,6 +39,8 @@
     "giphy-js-sdk-core": "^1.0.3",
     "intersection-observer": "^0.5.0",
     "js-yaml": "^3.12.0",
+    "react": "^16.4.1",
+    "react-dom": "^16.4.1",
     "react-helmet": "^5.2.0",
     "react-markdown": "^3.3.2",
     "react-transition-group": "^2.3.1",
@@ -55,7 +57,7 @@
     "gatsby-transformer-json": "^1.0.19",
     "ora": "^2.1.0",
     "prepend": "^1.0.2",
-    "prettier": "^1.13.4",
+    "prettier": "^1.13.5",
     "prettier-stylelint": "^0.4.2",
     "slugify": "^1.3.0",
     "stylelint": "^9.2.1",
