+++ src/pages/404.jsx
@@ -1,6 +1,7 @@
 import React, { Component } from 'react'
 import Link from 'gatsby-link'
 import giphyAPI from 'giphy-js-sdk-core'
+import Layout from '../components/Layout'
 import Content from '../components/atoms/Content'
 import './404.scss'
 
@@ -40,22 +41,24 @@ class NotFound extends Component {
 
   render() {
     return (
-      <Content className="content content--404">
-        <h1>Shenanigans, page not found.</h1>
-        <p>
-          You might want to check the url, or{' '}
-          <Link to={'/'}>go back to the homepage</Link>. Or just check out some
-          fail gifs, entirely your choice.
-        </p>
+      <Layout>
+        <Content className="content content--404">
+          <h1>Shenanigans, page not found.</h1>
+          <p>
+            You might want to check the url, or{' '}
+            <Link to={'/'}>go back to the homepage</Link>. Or just check out
+            some fail gifs, entirely your choice.
+          </p>
 
-        <video className="gif" src={this.state.gif} autoPlay loop />
+          <video className="gif" src={this.state.gif} autoPlay loop />
 
-        <div>
-          <a className="gif__action" href="#" onClick={this.handleClick}>
-            Show me another cat fail gif
-          </a>
-        </div>
-      </Content>
+          <div>
+            <a className="gif__action" href="#" onClick={this.handleClick}>
+              Show me another cat fail gif
+            </a>
+          </div>
+        </Content>
+      </Layout>
     )
   }
 }
