+++ src/pages/index.jsx
@@ -1,6 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import Link from 'gatsby-link'
+import Layout from '../components/Layout'
 import ProjectImage from '../components/atoms/ProjectImage'
 import FullWidth from '../components/atoms/FullWidth'
 import styles from './index.module.scss'
@@ -9,20 +10,22 @@ const Home = ({ data }) => {
   const projects = data.allProjectsYaml.edges
 
   return (
-    <FullWidth className="projects">
-      {projects.map(({ node }) => {
-        const { slug, title, img } = node
+    <Layout>
+      <FullWidth className="projects">
+        {projects.map(({ node }) => {
+          const { slug, title, img } = node
 
-        return (
-          <article className={styles.project} key={slug}>
-            <Link to={slug}>
-              <h1 className={styles.title}>{title}</h1>
-              <ProjectImage sizes={img.childImageSharp.sizes} alt={title} />
-            </Link>
-          </article>
-        )
-      })}
-    </FullWidth>
+          return (
+            <article className={styles.project} key={slug}>
+              <Link to={slug}>
+                <h1 className={styles.title}>{title}</h1>
+                <ProjectImage fluid={img.childImageSharp.sizes} alt={title} />
+              </Link>
+            </article>
+          )
+        })}
+      </FullWidth>
+    </Layout>
   )
 }
